requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0

# Web scraping
playwright>=1.40.0

# Database
# SQLite is included in Python standard library

# LLM/AI
# Using OpenAI-compatible API, no specific client needed

# Scheduling (future)
# apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality (optional)
# black>=23.0.0
# ruff>=0.1.0
